@@ -1,27 +1,35 @@
 {
   "name": "eslint-config-codex",
-  "version": "1.9.2",
+  "version": "2.0.0",
   "main": "index.js",
   "repository": "github:codex-team/eslint-config",
   "license": "MIT",
+  "type": "module",
   "scripts": {
-    "lint:test": "eslint test/*{.ts,.js}"
+    "test": "eslint test/*",
+    "inspect": "npx @eslint/config-inspector"
   },
   "peerDependencies": {
-    "eslint": ">= 5.3.0"
+    "eslint": ">= 9.0.0"
   },
   "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^6.2.0",
-    "@typescript-eslint/parser": "^6.2.0",
     "eslint-config-standard": "17.1.0",
-    "eslint-plugin-import": "2.28.0",
-    "eslint-plugin-jsdoc": "^46.4.5",
-    "eslint-plugin-n": "^16.0.1",
+    "eslint-flat-config-utils": "^0.2.3",
+    "eslint-plugin-import": "2.29.1",
+    "eslint-plugin-jsdoc": "^48.2.3",
+    "eslint-plugin-n": "^17.2.0",
     "eslint-plugin-promise": "6.1.1",
-    "eslint-plugin-standard": "5.0.0"
+    "eslint-plugin-standard": "5.0.0",
+    "globals": "^15.0.0",
+    "vue-eslint-parser": "^9.4.2"
   },
   "devDependencies": {
-    "eslint": "^8.23.1",
-    "typescript": "^5.1.6"
+    "@eslint/js": "^9.0.0",
+    "@stylistic/eslint-plugin": "^1.7.0",
+    "eslint": "^9.0.0",
+    "eslint-config-flat-gitignore": "^0.1.5",
+    "eslint-plugin-vue": "^9.25.0",
+    "typescript": "^5.4.5",
+    "typescript-eslint": "^7.6.0"
   }
 }
